+++ src/libmaxtouch/utilfuncs.h
@@ -39,4 +39,4 @@ void print_objs(void);
 
 void write_to_object(int obj_num);
 int read_object(uint16_t object_type, uint8_t instance, uint16_t offset, size_t count, bool format);
-int mxt_convert_hex(char *hex, unsigned char *databuf, uint8_t *count, unsigned int buf_size);
+int mxt_convert_hex(char *hex, unsigned char *databuf, uint16_t *count, unsigned int buf_size);
